// ide/gearbox-studio/src/common/generated/ExplanationGraph.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExplanationNode } from "./ExplanationNode";
import type { NodeId } from "./NodeId";
import type { ProvenanceEdge } from "./ProvenanceEdge";

/**
 * Why the product is the way it is.
 */
export type ExplanationGraph = { nodes: { [key in NodeId]: ExplanationNode }, edges: Array<ProvenanceEdge>, };